pyproject.toml
[project]
name = "rec-sys"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annoy>=1.17.3",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "beautifulsoup4==4.14.2",
    "black>=25.12.0",
    "bs4==0.0.2",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "datasets==4.4.1",
    "dill==0.4.0",
    "dvc[gdrive]>=3.65.0",
    "fastapi>=0.127.0",
    "fasttext>=0.9.3",
    "filelock==3.20.0",
    "fire>=0.7.1",
    "flake8>=7.3.0",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "gensim>=4.4.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub>=0.34.0,<1.0",
    "hydra-core>=1.3.2",
    "idna==3.11",
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
    "mlflow>=3.7.0",
    "multidict==6.7.0",
    "multiprocess==0.70.18",
    "numpy==2.3.5",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "packaging==25.0",
    "pandas==2.3.3",
    "pandera>=0.27.0",
    "polars==1.35.2",
    "polars-runtime-32==1.35.2",
    "pre-commit>=4.5.0",
    "propcache==0.4.1",
    "pyarrow==22.0.0",
    "pydantic>=2.12.5",
    "python-dateutil==2.9.0.post0",
    "pytorch-lightning>=2.6.0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "ruff>=0.14.7",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "torch-geometric>=2.7.0",
    "tqdm==4.67.1",
    "tritonclient[grpc]>=2.64.0",
    "typer-slim==0.20.0",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
]
